+++ packages/yourloops/components/footer-links.tsx
@@ -29,16 +29,20 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
 
-import { makeStyles, Theme } from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
+import Box from "@material-ui/core/Box";
 import Container from "@material-ui/core/Container";
-import Grid from "@material-ui/core/Grid";
 import Link from "@material-ui/core/Link";
+import { Theme, makeStyles } from "@material-ui/core/styles";
 import Tooltip from "@material-ui/core/Tooltip";
+import LanguageIcon from "@material-ui/icons/Language";
 
+import LanguageSelector from "../components/language-select";
 import diabeloopUrls from "../lib/diabeloop-url";
+import { useAuth } from "../lib/auth";
 import config from "../lib/config";
 import metrics from "../lib/metrics";
+import diabeloopLabel from "diabeloop-label.svg";
+import diabeloopLogo from "diabeloop-logo.svg";
 
 interface FooterLinksProps {
   atBottom?: boolean;
@@ -46,50 +50,169 @@ interface FooterLinksProps {
 
 const footerStyle = makeStyles((theme: Theme) => {
   return {
-    container: {
+    bySpan: {
+      paddingLeft: "12px",
+      paddingRight: "12px",
+      [theme.breakpoints.down("xs")]: {
+        paddingRight: "0",
+      },
+    },
+    centerBox: {
+      alignItems: "center",
       display: "flex",
-      fontSize: "small",
+      justifyContent: "center",
+      paddingLeft: "15px",
+      paddingRight: "15px",
+      [theme.breakpoints.up("sm")]: {
+        flexWrap: "wrap",
+      },
+      [theme.breakpoints.down("sm")]: {
+        flexWrap: "wrap",
+        order: 1,
+        textAlign: "center",
+        width: "100%",
+      },
+      [theme.breakpoints.down("xs")]: {
+        justifyContent: "space-around",
+        marginLeft: "10px",
+        marginRight: "10px",
+      },
     },
-    containerAtBottom: {
+    container: {
+      alignItems: "center",
+      boxShadow: "0px -1px 4px rgba(0, 0, 0, 0.25)",
+      color: theme.palette.grey[700],
+      display: "flex",
+      fontSize: "12px",
+      padding: "11px",
+      [theme.breakpoints.down("sm")]: {
+        flexWrap: "wrap",
+      },
+    },
+    containerAbsolute: {
+      ["@media (min-height:630px)"]: {
+        bottom: "0",
+        position: "absolute",
+      },
+      height: "68px",
+      [theme.breakpoints.down("sm")]: {
+        height: "149px",
+      },
+      [theme.breakpoints.down("xs")]: {
+        height: "189px",
+      },
+    },
+    containerAuto: {
+      height: "68px",
       marginTop: "auto",
-      paddingBottom: "2em",
-      paddingTop: "2em",
+      paddingBottom: "11px",
+      paddingTop: "11px",
     },
-    rightLink: {
-      padding: theme.spacing(0.5), // eslint-disable-line no-magic-numbers
-      textAlign: "start",
-      marginBottom: "auto",
-      marginTop: "auto",
+    cookiesManagement: {
+      "&:hover": {
+        cursor: "pointer",
+      },
     },
-    centeredLink: {
-      padding: theme.spacing(0.5), // eslint-disable-line no-magic-numbers
-      textAlign: "center",
-      marginBottom: "auto",
-      marginTop: "auto",
-      overflow: "hidden",
-      whiteSpace: "nowrap",
+    diabeloopLogo: {
+      paddingRight: "3px",
     },
-    leftLink: {
-      padding: theme.spacing(0.5), // eslint-disable-line no-magic-numbers
-      textAlign: "end",
-      marginBottom: "auto",
-      marginTop: "auto",
+    diabeloopLink: {
+      [theme.breakpoints.down("xs")]: {
+        marginTop: "12px",
+      },
     },
-    selection: {
-      padding: theme.spacing(2),
-      textAlign: "center",
+    languageBox: {
+      display: "flex",
+      height: "20px",
+      alignItems: "center",
+      [theme.breakpoints.down("sm")]: {
+        width: "100%",
+        display: "inline-block",
+        marginTop: "10px",
+        marginBottom: "17px",
+      },
+      [theme.breakpoints.down("xs")]: {
+        marginBottom: "27px",
+        marginTop: "6px",
+      },
     },
-    cookiesButton: {
-      textTransform: "initial",
-      fontSize: "small",
+    languageIcon: {
+      width: "20px",
+      marginRight: "18px",
+      alignSelf: "center",
+      [theme.breakpoints.down("sm")]: {
+        marginTop: "3px",
+      },
+    },
+    languageSeparator: {
+      alignSelf: "center",
+      [theme.breakpoints.down("sm")]: {
+        display: "none",
+        visibility: "hidden",
+      },
+    },
+    leftBox: {
+      width: "134px",
+      [theme.breakpoints.down("sm")]: {
+        order: 2,
+      },
+    },
+    link: {
+      color: theme.palette.grey[700],
+      fontWeight: 400,
+      [theme.breakpoints.down("xs")]: {
+        marginBottom: "15px",
+        marginLeft: "0.5rem",
+        marginRight: "0.5rem",
+        textAlign: "center",
+      },
+    },
+    rightBox: {
+      display: "flex",
+      justifyContent: "right",
+      [theme.breakpoints.down("sm")]: {
+        order: 3,
+      },
+      [theme.breakpoints.down("xs")]: {
+        display: "flex",
+        flexDirection: "column",
+        textAlign: "right",
+      },
+    },
+    separator: {
+      paddingLeft: "15px",
+      paddingRight: "15px",
+      [theme.breakpoints.down("xs")]: {
+        display: "none",
+        visibility: "hidden",
+      },
+    },
+    sideBox: {
+      flex: "1",
+    },
+    supportButton: {
+      height: "46px",
+      width: "134px",
+      [theme.breakpoints.down("xs")]: {
+        marginTop: "10px",
+      },
+    },
+    svg: {
+      height: "12px",
+      verticalAlign: "middle",
+      display: "inline-block",
+    },
+    versionSpan: {
+      textDecoration: "underline",
     },
   };
 }, { name: "footer-component-styles" });
 
 function FooterLinks(props: FooterLinksProps): JSX.Element {
   const { t, i18n } = useTranslation("yourloops");
+  const { user } = useAuth();
   const classes = footerStyle();
-  const containerClassName = props.atBottom ? `${classes.container} ${classes.containerAtBottom}` : classes.container;
+  const containerClassName = props.atBottom ? classes.containerAuto : classes.containerAbsolute;
 
   const handleShowCookieBanner = () => {
     if (typeof window.openAxeptioCookies === "function") {
@@ -102,33 +225,97 @@ function FooterLinks(props: FooterLinksProps): JSX.Element {
   };
 
   return (
-    <Container id="footer-links-container" className={containerClassName} maxWidth="sm">
-      <Grid id="footer-links" container>
-        <Grid item xs={4} className={classes.rightLink}>
-          <Link id="footer-link-url-privacy-policy" target="_blank" href={diabeloopUrls.getPrivacyPolicyUrL(i18n.language)} rel="nofollow" onClick={metricsPdfDocument("privacy_policy")}>{t("privacy-policy")}</Link>
-        </Grid>
-        <Grid id="footer-link-app-name" item xs={4} className={classes.centeredLink}>
-          <Tooltip id="footer-link-tooltip-app-release-notes" title={t("tooltip-release-notes") as string} aria-label={t("tooltip-release-notes")} placement="right-start">
-            <Link id="footer-link-url-release-notes" target="_blank" href={diabeloopUrls.getReleaseNotesURL()} rel="nofollow" onClick={metricsPdfDocument("release_notes")}>
-              {`${t("brand-name")} ${config.VERSION}`}
+    <Container id="footer-links-container" className={`${classes.container} ${containerClassName}`} maxWidth={false}>
+      <Box className={`${classes.sideBox} ${classes.leftBox}`}>
+        <Box className={classes.supportButton}></Box>
+      </Box>
+      <Box className={classes.centerBox}>
+        {!user &&
+          <Box id="footer-language-box" className={classes.languageBox}>
+            <LanguageIcon className={classes.languageIcon}></LanguageIcon>
+            <LanguageSelector />
+            <Box className={`${classes.separator} ${classes.languageSeparator}`}>|</Box>
+          </Box>
+        }
+        <Link
+          id="footer-link-url-privacy-policy"
+          target="_blank"
+          href={diabeloopUrls.getPrivacyPolicyUrL(i18n.language)}
+          rel="nofollow"
+          onClick={metricsPdfDocument("privacy_policy")}
+          className={classes.link}
+        >
+          {t("privacy-policy")}
+        </Link>
+        <Box className={classes.separator}>|</Box>
+        <Link
+          id="footer-link-url-terms"
+          target="_blank"
+          href={diabeloopUrls.getTermsUrL(i18n.language)}
+          rel="nofollow"
+          onClick={metricsPdfDocument("terms")}
+          className={classes.link}
+        >
+          {t("terms-of-use")}
+        </Link>
+        <Box className={classes.separator}>|</Box>
+        <Link
+          id="footer-link-url-intended-use"
+          target="_blank"
+          href={diabeloopUrls.getIntendedUseUrL(i18n.language)}
+          rel="nofollow"
+          onClick={metricsPdfDocument("intended_use")}
+          className={classes.link}
+        >
+          {t("intended-use")}
+        </Link>
+        <Box className={classes.separator}>|</Box>
+        <Link id="footer-link-cookies-management"
+          className={`${classes.link} ${classes.cookiesManagement}`}
+          onClick={handleShowCookieBanner}
+        >
+          {t("cookies-management")}
+        </Link>
+        <Box className={classes.separator}>|</Box>
+        <Link
+          id="footer-link-url-cookies-policy"
+          target="_blank"
+          href={diabeloopUrls.getCookiesPolicyUrl(i18n.language)}
+          rel="nofollow"
+          onClick={metricsPdfDocument("yourloops-cookiepolicy")}
+          className={classes.link}
+        >
+          {t("cookies-policy")}
+        </Link>
+      </Box>
+      <Box className={`${classes.sideBox} ${classes.rightBox}`}>
+        <Box>
+          {t("brand-name")}
+          <Tooltip
+            id="footer-link-tooltip-app-release-notes"
+            title={t("tooltip-release-notes") as string}
+            aria-label={t("tooltip-release-notes")}
+            placement="right-start"
+          >
+            <Link
+              id="footer-link-url-release-notes"
+              target="_blank"
+              href={diabeloopUrls.getReleaseNotesURL()}
+              rel="nofollow"
+              onClick={metricsPdfDocument("release_notes")}
+              className={classes.link}
+            >
+              <span className={classes.versionSpan}>&nbsp;v{config.VERSION}</span>
             </Link>
           </Tooltip>
-        </Grid>
-        <Grid item xs={4} className={classes.leftLink}>
-          <Link id="footer-link-url-support" target="_blank" href={diabeloopUrls.SupportUrl} rel="nofollow" onClick={() => metrics.send("support", "click_url_support")}>{t("footer-link-url-support")}</Link>
-        </Grid>
-        <Grid item xs={4} className={classes.rightLink}>
-          <Link id="footer-link-url-terms" target="_blank" href={diabeloopUrls.getTermsUrL(i18n.language)} rel="nofollow" onClick={metricsPdfDocument("terms")}>{t("terms-of-use")}</Link>
-        </Grid>
-        <Grid item xs={4} className={classes.centeredLink}>
-          {/* TODO: Add tooltip + aria label */}
-          <Button id="footer-link-cookies" color="primary" variant="text" className={classes.cookiesButton} onClick={handleShowCookieBanner}>{t("cookies")}</Button>
-        </Grid>
-        <Grid item xs={4} className={classes.leftLink}>
-          <Link id="footer-link-url-intended-use" target="_blank" href={diabeloopUrls.getIntendedUseUrL(i18n.language)} rel="nofollow" onClick={metricsPdfDocument("intended_use")}>{t("intended-use")}</Link>
-        </Grid>
-      </Grid>
-    </Container>
+          <span className={classes.bySpan}>by </span>
+        </Box>
+        <Link id="footer-link-url-diabeloop" className={classes.diabeloopLink} target="_blank" href={diabeloopUrls.SupportUrl} rel="nofollow" >
+          <img src={diabeloopLogo} alt={t("alt-img-logo")} className={`${classes.svg} ${classes.diabeloopLogo}`} />
+          <img src={diabeloopLabel} alt={t("alt-img-logo")} className={classes.svg} />
+        </Link>
+      </Box>
+    </Container >
   );
 }