+++ packages/yourloops/lib/diabeloop-url.tsx
@@ -38,6 +38,7 @@ class DiabeloopUrl {
   private termsUrl: string;
   private privacyPolicyUrl: string;
   private intendedUseUrL: string;
+  private cookiesPolicyUrL: string;
   private supportUrL: string;
 
   constructor() {
@@ -45,6 +46,7 @@ class DiabeloopUrl {
     this.termsUrl = `${this.rootPathName}terms.pdf`;
     this.privacyPolicyUrl = `${this.rootPathName}data-privacy.pdf`;
     this.intendedUseUrL = `${this.rootPathName}intended-use.pdf`;
+    this.cookiesPolicyUrL = `${this.rootPathName}yourloops-cookiepolicy.pdf`;
     this.supportUrL = "https://www.diabeloop.com";
   }
 
@@ -84,6 +86,11 @@ class DiabeloopUrl {
   getIntendedUseUrL(currentLangue: string): string {
     this.intendedUseUrL = `${this.rootPathName}intended-use.${currentLangue}.pdf`;
     return this.intendedUseUrL;
+  }
+
+  getCookiesPolicyUrl(currentLangue: string): string {
+    this.cookiesPolicyUrL = `${this.rootPathName}yourloops-cookiepolicy.${currentLangue}.pdf`;
+    return this.cookiesPolicyUrL;
   }
 
   getReleaseNotesURL(): string {