+++ packages/yourloops/pages/password-reset/confirm-password-reset.tsx
@@ -36,7 +36,6 @@ import Container from "@material-ui/core/Container";
 import Grid from "@material-ui/core/Grid";
 
 import brandingLogo from "branding/logo.png";
-import LanguageSelector from "../../components/language-select";
 import ResetPasswordContent from "./reset-password-content";
 
 const formStyle = makeStyles((theme: Theme) => {
@@ -44,11 +43,6 @@ const formStyle = makeStyles((theme: Theme) => {
     mainContainer: {
       display: "flex",
       flexDirection: "column",
-    },
-    divLanguageSelector: {
-      marginLeft: "auto",
-      marginRight: "auto",
-      padding: "16px",
     },
     Card: {
       display: "flex",
@@ -98,9 +92,6 @@ function ConfirmPasswordResetPage(): JSX.Element {
           </Card>
         </Grid>
       </Grid>
-      <div className={classes.divLanguageSelector}>
-        <LanguageSelector />
-      </div>
     </Container>
   );
 }