+++ packages/yourloops/test/components/index.ts
@@ -28,16 +28,19 @@
 
 import testDatePickers from "./date-pickers";
 import testSnackbar from "./utils/snackbar.test";
-import TestPasswordStrengthMeter from "./password-strength-meter.test";
+import testFooterLink from "./footer-links.test";
+import testPasswordStrengthMeter from "./password-strength-meter.test";
 import testSwitchRole from "./switch-role";
 import testDropdowns from "./dropdown";
 
 function testComponents(): void {
   describe("DatePickers", testDatePickers);
+  describe("Footer", testFooterLink);
   describe("Snackbar", testSnackbar);
-  describe("Dropdowns", testDropdowns);
+  describe("Footer", testFooterLink);
+  describe("Password strength meter", testPasswordStrengthMeter);
   describe("SwitchRole", testSwitchRole);
-  describe("Password strength meter", TestPasswordStrengthMeter);
+  describe("Dropdowns", testDropdowns);
 }
 
 export default testComponents;